Fix ‘NodeVisitor’ cannot be named TypeScript error
Something seems to be going a bit spoopy with TypeScript’s type inference here. When trying to write the declarations for the tokenizer-event-handlers.ts file, it isn’t able to infer the
type of the exported `traverse` member in a way that allows it to print the `NodeVisitor` argument
type.

Defining an interface for the exported object seems to work. I’m not sure why `typeof traverse` works when the inferred type doesn’t, but this fixes an issue where people would see a type error when importing `@glimmer/compiler`.